requirements.txt
anthropic>=0.25.0
jinja2>=3.1.0
playwright>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
